Mock react navigation in __mocks__
